requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
Authlib==1.6.8
bandit==1.9.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.4
cron-descriptor==1.4.5
cryptography==46.0.5
Django==6.0.1
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dparse==0.6.4
filelock==3.24.3
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mccabe==0.7.0
mdurl==0.1.2
nltk==3.9.2
packaging==26.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.2.1
regex==2026.2.19
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
stevedore==5.7.0
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
gunicorn==21.2.0
